refactor(backend): better modularize backends
