--- examples/demo-canvas/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "demo"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-clap = { version = "4.5.23", features = ["derive"] }
-rand = { version = "0.8.5", features = ["small_rng"], default-features = false }
-ratzilla = { path = "../../" }
