+++ src/lib.rs
@@ -1,16 +1,12 @@
 mod backend;
-mod canvas_backend;
-mod canvas_utils;
 mod render;
-mod render_canvas;
 pub mod utils;
 
 pub mod widgets;
 
-pub use backend::DomBackend;
-pub use canvas_backend::WasmCanvasBackend;
+pub use backend::canvas::CanvasBackend;
+pub use backend::dom::DomBackend;
 pub use render::RenderOnWeb;
-pub use render_canvas::RenderOnWebCanvas;
 
 // Re-export ratatui crate.
 pub use ratatui;
